introduction
This is a demo application to register the member from school. It is based on Power apps and Power automate, and comparing to html and css, it is very easy to design not only for cloud services but for mobile applications as well. The demo program can check the information of students from different courses and add new information. By using Power automate, an email address can be related and get an confirmation email whenever the information of the app is updated.